+++ test/support/pet.ex
@@ -24,7 +24,7 @@ defmodule MyApp.Pet do
       :custom,
       :reverse_name
     ],
-    sortable: [:name, :age, :owner_name, :owner_age],
+    sortable: [:name, :age, :owner_name, :owner_age, :name_length],
     max_limit: 1000,
     adapter_opts: [
       compound_fields: [
@@ -58,6 +58,10 @@ defmodule MyApp.Pet do
         reverse_name: [
           filter: {__MODULE__, :reverse_name_filter, []},
           ecto_type: :string
+        ],
+        name_length: [
+          sorter: {__MODULE__, :name_length_sorter, [source: :name]},
+          ecto_type: :integer
         ]
       ]
     ]
@@ -90,6 +94,14 @@ defmodule MyApp.Pet do
   def reverse_name_filter(query, %Flop.Filter{value: value}, _) do
     reversed = value
     where(query, [p], p.name == ^reversed)
+  end
+
+  def name_length_sorter(query, direction, opts) do
+    source = Keyword.fetch!(opts, :source)
+    send(self(), {:sort, direction, opts})
+    order_by(query, [p], [
+      {^direction, fragment("length(?)", field(p, ^source))}
+    ])
   end
 
   def get_field(%__MODULE__{owner: %Owner{age: age}}, :owner_age), do: age